requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
httpx==0.28.1
jinja2==3.1.5
python-multipart==0.0.20
pyyaml==6.0.2
